requirements.txt
## Python Packages
pip
setuptools
raven
pytz
mysqlclient
coverage
selenium
vobject
pil
#whitenoise
#invoke
#pip-tools
#cookiecutter
#pytest
#sh
#elasticsearch
#logstash
#requests
#ipython
#ipdb
#flake8
#tablib
#WebTest
## Django Packages
django>=2.0.0,<2.1.0
django-tables2
django-filter
django-widget-tweaks
django-digg-paginator
django-debug-toolbar
django-configurations
django-simple-history
django-crispy-forms
django-extra-views
django-model-utils
django-taggit
django-mptt
django-extensions
#django-vanilla-views
#django-webtest
#model-mommy
#django-haystack
#django-redis
#django-clear-cache
#elasticsearch_dsl
#gunicorn